# swagger.yaml
openapi: 3.0.0
paths:
...
  /policies/{policyId}/tag/{tagName}:
    get:
      operationId: PolicyApi_getBlockByTagName
      summary: Requests block data.
      description: >-
        Requests block data by tag. Only users with a role that described in
        block are allowed to make the request.
      parameters: []
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorDTO'
      tags: *ref_4
      security:
        - bearerAuth: []
        - bearerAuth: []
...
  /schemas:
    get:
      tags:
        - schemas
      description: Returns all schemas.
      security:
        - bearerAuth: []
      summary: Returns all schemas.
      parameters:
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: >-
            The number of pages to skip before starting to collect the result
            set
          examples:
            pageIndex:
              summary: Example of a pageIndex
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
          examples:
            pageSize:
              summary: Example of a pageSize
              value: 100
      responses:
        '200':
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schemas/{topicId}:
    post:
      tags:
        - schemas
      description: >-
        Creates new schema. Only users with the Standard Registry role are
        allowed to make the request.
      parameters:
        - in: path
          name: topicId
          schema:
            type: string
          required: true
          description: Topic ID.
          examples:
            topicId:
              summary: Example of a Topic ID
              value: 0.0.00000001
      security:
        - bearerAuth: []
      summary: Create new schema.
      requestBody:
        description: Object that contains a valid schema.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schema'
      responses:
        '201':
          description: Created.
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - schemas
      description: Returns all schemas by topicId.
      security:
        - bearerAuth: []
      summary: Returns all schemas by topicId.
      parameters:
        - in: path
          name: topicId
          schema:
            type: string
          required: true
          description: Topic ID.
          examples:
            topicId:
              summary: Example of a Topic ID
              value: 0.0.00000001
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: >-
            The number of pages to skip before starting to collect the result
            set
          examples:
            pageIndex:
              summary: Example of a pageIndex
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
          examples:
            pageSize:
              summary: Example of a pageSize
              value: 100
      responses:
        '200':
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /schemas/{schemaId}:
    put:
      tags:
        - schemas
      description: >-
        Updates the schema with the provided schema ID. Only users with the
        Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: schemaId
          schema:
            type: string
          required: true
          description: Schema ID.
      security:
        - bearerAuth: []
      summary: Updates the schema.
      requestBody:
        description: Object that contains a valid schema.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schema'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - schemas
      description: >-
        Deletes the schema with the provided schema ID. Only users with the
        Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: schemaId
          schema:
            type: string
          required: true
          description: Schema ID.
      security:
        - bearerAuth: []
      summary: Deletes the schema.
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schemas/{schemaId}/publish:
    put:
      tags:
        - schemas
      description: >-
        Publishes the schema with the provided (internal) schema ID onto IPFS,
        sends a message featuring IPFS CID into the corresponding Hedera topic.
        Only users with the Standard Registry role are allowed to make the
        request.
      parameters:
        - in: path
          name: schemaId
          schema:
            type: string
          required: true
          description: Schema ID.
      security:
        - bearerAuth: []
      summary: Publishes the schema.
      requestBody:
        description: Object that contains policy version.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                version:
                  type: string
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /tags/search:
    post:
      tags:
        - tags
      description: Search tags.
      security:
        - bearerAuth: []
      requestBody:
        description: Object that contains filters.
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - entity
                    - target
                  properties:
                    entity:
                      type: string
                      enum:
                        - Schema
                        - Policy
                        - Token
                        - Module
                        - Contract
                        - PolicyDocument
                    target:
                      type: string
                - type: object
                  required:
                    - entity
                    - targets
                  properties:
                    entity:
                      type: string
                      enum:
                        - Schema
                        - Policy
                        - Token
                        - Module
                        - Contract
                        - PolicyDocument
                    targets:
                      type: array
                      items:
                        type: string
            examples:
              Single:
                value:
                  entity: PolicyDocument
                  target: targetId1
              Multiple:
                value:
                  entity: PolicyDocument
                  targets:
                    - targetId1
                    - targetId2
      summary: Search tags.
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                description: a (targetId, Tags) map. `targetId1` is an example key
                properties:
                  targetId1:
                    $ref: '#/components/schemas/TagMap'
                additionalProperties:
                  $ref: '#/components/schemas/TagMap'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tags/{uuid}:
    delete:
      tags:
        - tags
      description: Delete tag.
      security:
        - bearerAuth: []
      summary: Delete tag.
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            example: 00000000-0000-0000-0000-000000000000
          required: true
          description: Tag identifier
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tags/synchronization:
    post:
      tags:
        - tags
      description: synchronization.
      security:
        - bearerAuth: []
      requestBody:
        description: Object that contains filters.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - entity
                - target
              properties:
                entity:
                  type: string
                  enum:
                    - Schema
                    - Policy
                    - Token
                    - Module
                    - Contract
                    - PolicyDocument
                  example: PolicyDocument
                target:
                  type: string
                  example: targetId
      summary: synchronization.
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagMap'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tags/schemas:
    get:
      tags:
        - tags
      description: Returns all schema.
      security:
        - bearerAuth: []
      summary: Returns all schemas.
      parameters:
        - in: query
          name: pageIndex
          schema:
            type: integer
          description: >-
            The number of pages to skip before starting to collect the result
            set
          examples:
            pageIndex:
              summary: Example of a pageIndex
              value: 0
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The numbers of items to return
          examples:
            pageSize:
              summary: Example of a pageSize
              value: 100
      responses:
        '200':
          description: Successful operation.
          headers:
            x-total-count:
              schema:
                type: integer
              description: Total items in the collection.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - tags
      description: >-
        Creates new schema. Only users with the Standard Registry role are
        allowed to make the request.
      security:
        - bearerAuth: []
      summary: Creates new schema.
      requestBody:
        description: Object that contains a valid schema.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schema'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tags/schemas/{schemaId}:
    delete:
      tags:
        - tags
      description: >-
        Deletes the schema with the provided schema ID. Only users with the
        Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Delete the schema.
      parameters:
        - in: path
          name: schemaId
          schema:
            type: string
          required: true
          description: Schema ID.
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - tags
      description: >-
        Updates the schema with the provided schema ID. Only users with the
        Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: schemaId
          schema:
            type: string
          required: true
          description: Schema ID.
      security:
        - bearerAuth: []
      summary: Updates the schema.
      requestBody:
        description: Object that contains a valid schema.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schema'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tags/schemas/{schemaId}/publish:
    put:
      tags:
        - tags
      description: >-
        Publishes the schema with the provided (internal) schema ID onto IPFS,
        sends a message featuring IPFS CID into the corresponding Hedera topic.
        Only users with the Standard Registry role are allowed to make the
        request.
      parameters:
        - in: path
          name: schemaId
          schema:
            type: string
          required: true
          description: Schema ID.
      security:
        - bearerAuth: []
      summary: Publishes the schema.
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tags/schemas/published:
    get:
      tags:
        - tags
      description: Return a list of all published schemas.
      security:
        - bearerAuth: []
      summary: Return a list of all published schemas.
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schema'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /policies/{policyId}/tag/{tag}:
    get:
      tags:
        - policies
      description: >-
        Requests block ID from a policy by tag. Only users with the Standard
        Registry and Installer roles are allowed to make the request.
      parameters:
        - in: path
          name: policyId
          schema:
            type: string
          required: true
          description: Selected policy ID.
        - in: path
          name: tag
          schema:
            type: string
          required: true
          description: Tag from the selected policy.
      summary: Requests block ID from a policy by tag.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /tags/:
    post:
      tags:
        - tags
      description: Creates new tag.
      security:
        - bearerAuth: []
      summary: Creates new tag.
      requestBody:
        description: Object that contains tag information.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized.
        '403':
          description: Forbidden.
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
info:
  title: Guardian
  description: >-
    The Guardian is a modular open-source solution that includes best-in-class
    identity management and decentralized ledger technology (DLT) libraries. At
    the heart of the Guardian solution is a sophisticated Policy Workflow Engine
    (PWE) that enables applications to offer a requirements-based tokenization
    implementation.
  version: 2.17.0
  contact:
    name: API developer
    url: https://envisionblockchain.com
    email: info@envisionblockchain.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: /api/v1
    description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InternalServerErrorDTO:
      type: object
      properties:
        code:
          type: number
        message:
          type: string
      required:
        - code
        - message
...
    Schema:
      type: object
      required:
        - id
        - name
        - description
        - entity
        - document
      properties:
        id:
          type: string
        iri:
          type: string
        uuid:
          type: string
        name:
          type: string
        description:
          type: string
        entity:
          type: string
        hash:
          type: string
        status:
          type: string
        document:
          oneOf:
            - type: string
            - type: object
        topicId:
          type: string
        version:
          type: string
        owner:
          type: string
        messageId:
          type: string
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    TagMap:
      type: object
      required:
        - entity
        - target
        - refreshDate
        - tags
      properties:
        entity:
          type: string
          enum:
            - Schema
            - Policy
            - Token
            - Module
            - Contract
            - PolicyDocument
          example: PolicyDocument
        target:
          type: string
          example: db id
        refreshDate:
          type: string
          example: '1900-01-01T00:00:00.000Z'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
...